adding relation user and person

+++ src/main/java/com/postech/techchallengefase1/domain/person/service/PersonService.java
@@ -2,12 +2,16 @@ package com.postech.techchallengefase1.domain.person.service;
 
 import com.postech.techchallengefase1.domain.exception.ApiException;
 import com.postech.techchallengefase1.domain.person.dto.CreatePersonDTO;
+import com.postech.techchallengefase1.domain.person.dto.PersonWithUserDTO;
 import com.postech.techchallengefase1.domain.person.dto.UpdatePersonDTO;
 import com.postech.techchallengefase1.domain.person.entity.Person;
+import com.postech.techchallengefase1.domain.person.enuns.Gender;
+import com.postech.techchallengefase1.domain.person.enuns.Relationship;
 import com.postech.techchallengefase1.domain.person.repository.PersonRepository;
 import com.postech.techchallengefase1.domain.user.entity.User;
 import com.postech.techchallengefase1.domain.user.repository.UserRepository;
 import lombok.AllArgsConstructor;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
@@ -20,27 +24,116 @@ public class PersonService {
 
     private final UserRepository userRepository;
 
-    public Person savePerson(CreatePersonDTO person, String username) {
+    private void checkParentRelationship(User user) {
+        if (user.getPersons().stream().filter(person -> Relationship.PARENT.equals(person.getRelationship())).count() >= 2) {
+            throw new ApiException("Cannot create more than 2 persons with PARENT relationship",
+                    HttpStatus.BAD_REQUEST.value());
+        }
+    }
+
+    private void checkPartnerRelationship(User user) {
+        if (user.getPersons().stream().anyMatch(person -> Relationship.PARTNER.equals(person.getRelationship()))) {
+            throw new ApiException("Cannot create more than 1 person with PARTNER relationship",
+                    HttpStatus.BAD_REQUEST.value());
+        }
+    }
+
+    private void checkSpouseRelationship(User user) {
+        System.out.println(user.getPersons().stream().filter(person -> Relationship.SPOUSE.equals(person.getRelationship())).count());
+        if (user.getPersons().stream().anyMatch(person -> Relationship.SPOUSE.equals(person.getRelationship()))) {
+            throw new ApiException("Cannot create more than 1 person with SPOUSE relationship",
+                    HttpStatus.BAD_REQUEST.value());
+        }
+    }
+
+    public PersonWithUserDTO savePerson(CreatePersonDTO createPersonDTO, String username) {
         User user = userRepository.findByUsername(username).orElseThrow(() ->
                 new ApiException("User not found", HttpStatus.NOT_FOUND.value()));
 
-        Person personToSave = person.getPerson();
-        personToSave.setUser(user);
-
-        Person newPerson = personRepository.save(personToSave);
-        if (newPerson.getId() == null) {
-            throw new ApiException("Person already exists", HttpStatus.CONFLICT.value());
+        if (Relationship.OWNER.equals(createPersonDTO.getRelationship())) {
+            throw new ApiException("Cannot create person with OWNER relationship",
+                    HttpStatus.BAD_REQUEST.value());
         }
-        return newPerson;
+
+        switch (createPersonDTO.getRelationship()) {
+            case PARENT -> checkParentRelationship(user);
+            case PARTNER -> checkPartnerRelationship(user);
+            case SPOUSE -> checkSpouseRelationship(user);
+        }
+
+        Person person = new Person();
+        person.setName(createPersonDTO.getName());
+        person.setCpf(createPersonDTO.getCpf());
+        person.setDateOfBirth(createPersonDTO.getDateOfBirth());
+        person.setGender(createPersonDTO.getGender());
+        person.setRelationship(createPersonDTO.getRelationship());
+        person.setUser(user);
+
+        try {
+            personRepository.save(person);
+            return PersonWithUserDTO.toDTO(person);
+        } catch (DataIntegrityViolationException e) {
+            throw new ApiException("CPF already registered", HttpStatus.CONFLICT.value());
+        }
+
     }
 
-    public List<Person> getAllPerson() {
-        return personRepository.findAll();
+    public List<PersonWithUserDTO> getAllPerson() {
+        List<Person> persons = personRepository.findAll();
+
+        if (persons.isEmpty()) {
+            throw new ApiException("No person found", HttpStatus.NOT_FOUND.value());
+        }
+
+        return persons.stream().map(PersonWithUserDTO::toDTO).toList();
     }
 
-    public Person getPersonById(Long id) {
-        return personRepository.findById(id).orElseThrow(() ->
+    public List<PersonWithUserDTO> getPersonByName(String name) {
+        List<Person> persons = personRepository.findByName(name).orElseThrow(() ->
+                new ApiException("No person found", HttpStatus.NOT_FOUND.value()));
+
+        return persons.stream().map(PersonWithUserDTO::toDTO).toList();
+
+    }
+
+    public List<PersonWithUserDTO> getPersonByRelationship(String relationship) {
+        try {
+            Relationship convertedRelationship = Relationship.valueOf(relationship.toUpperCase());
+            List<Person> persons = personRepository.findByRelationship(convertedRelationship).orElseThrow(() ->
+                    new ApiException("No person found", HttpStatus.NOT_FOUND.value()));
+
+            return persons.stream().map(PersonWithUserDTO::toDTO).toList();
+        } catch (IllegalArgumentException e) {
+            throw new ApiException("Invalid relationship", HttpStatus.BAD_REQUEST.value());
+        }
+    }
+
+    public List<PersonWithUserDTO> getPersonByGender(String gender) {
+        try {
+            Gender convertedGender = Gender.valueOf(gender.toUpperCase());
+            List<Person> persons = personRepository.findByGender(convertedGender).orElseThrow(() ->
+                    new ApiException("No person found", HttpStatus.NOT_FOUND.value()));
+
+            return persons.stream().map(PersonWithUserDTO::toDTO).toList();
+        } catch (IllegalArgumentException e) {
+            throw new ApiException("Invalid gender", HttpStatus.BAD_REQUEST.value());
+        }
+    }
+
+    public List<PersonWithUserDTO> getPersonByCpf(String cpf) {
+
+        List<Person> persons = personRepository.findByCpf(cpf).orElseThrow(() ->
+                new ApiException("No person found", HttpStatus.NOT_FOUND.value()));
+
+        return persons.stream().map(PersonWithUserDTO::toDTO).toList();
+
+    }
+
+    public PersonWithUserDTO getPersonById(Long id) {
+        Person person = personRepository.findById(id).orElseThrow(() ->
                 new ApiException("Person not found", HttpStatus.NOT_FOUND.value()));
+
+        return PersonWithUserDTO.toDTO(person);
     }
 
     public void deletePersonById(Long id) {
@@ -48,21 +141,31 @@ public class PersonService {
             throw new ApiException("Person not found", HttpStatus.NOT_FOUND.value());
         }
 
+        if (Relationship.OWNER.equals(personRepository.findById(id).get().getRelationship())) {
+            throw new ApiException("Cannot delete person with OWNER relationship",
+                    HttpStatus.BAD_REQUEST.value());
+        }
+
         personRepository.deleteById(id);
     }
 
-    public Person updatePerson(UpdatePersonDTO person) {
-        Person personToUpdate = personRepository.findById(person.getId()).orElseThrow(() ->
+    public PersonWithUserDTO updatePerson(UpdatePersonDTO person, Long id) {
+        if(person.getRelationship() != null && Relationship.OWNER.equals(person.getRelationship())) {
+            throw new ApiException("Cannot update person with OWNER relationship",
+                    HttpStatus.BAD_REQUEST.value());
+        }
+
+        Person personToUpdate = personRepository.findById(id).orElseThrow(() ->
                 new ApiException("Person not found", HttpStatus.NOT_FOUND.value()));
 
-        Person updatedPerson = new Person();
-        updatedPerson.setId(personToUpdate.getId());
-        updatedPerson.setName(person.getName() != null ? person.getName() : personToUpdate.getName());
-        updatedPerson.setDateOfBirth(person.getDateOfBirth() != null ? person.getDateOfBirth() : personToUpdate.getDateOfBirth());
-        updatedPerson.setGender(person.getGender() != null ? person.getGender() : personToUpdate.getGender());
-        updatedPerson.setRelationship(person.getRelationship() != null ? person.getRelationship() : personToUpdate.getRelationship());
+        personToUpdate.setName(person.getName() != null ? person.getName() : personToUpdate.getName());
+        personToUpdate.setDateOfBirth(person.getDateOfBirth() != null ? person.getDateOfBirth() : personToUpdate.getDateOfBirth());
+        personToUpdate.setGender(person.getGender() != null ? person.getGender() : personToUpdate.getGender());
+        personToUpdate.setRelationship(person.getRelationship() != null ? person.getRelationship() : personToUpdate.getRelationship());
 
-        return personRepository.save(updatedPerson);
+        personRepository.save(personToUpdate);
+
+        return PersonWithUserDTO.toDTO(personToUpdate);
     }
 
 }
